+++ src/test/java/gmibank/test-data/countryId.txt
@@ -118,7 +118,7 @@
 3
 5
 24355
-1
+22307
 4
 1752
 18901
@@ -144,7 +144,6 @@
 18931
 18932
 18936
-22309
 18941
 18946
 22327
@@ -155,11 +154,12 @@
 18948
 18949
 18950
+1
 22306
 22303
 18937
-22307
 22304
+22309
 24381
 24382
 24383
@@ -779,7 +779,6 @@
 62133
 62134
 62135
-62136
 62137
 62138
 62139
@@ -851,3 +850,69 @@
 62708
 62716
 62717
+62719
+62720
+62721
+62722
+62723
+62724
+62725
+62726
+62727
+62728
+62729
+62730
+62731
+62732
+62733
+62734
+62735
+62736
+62737
+62738
+62739
+62740
+62741
+62742
+62743
+62744
+62745
+62746
+62747
+62748
+62749
+62750
+62851
+62852
+62853
+62854
+62855
+62856
+62857
+62858
+62859
+62860
+62861
+62862
+62863
+62864
+62865
+62866
+62867
+62868
+62869
+62136
+62870
+62871
+62872
+62873
+62874
+62875
+62876
+62877
+62879
+62880
+62881
+62882
+62883
+62885
